Merged the System.Startup repository
Merged the Wizardry System.Startup repository into the Wizardry.Code repository since there are many crossovers when working on the Wizardry Recompiled project. Often updates happen simultaneously on both the boot and scenarios diskettes images of Wizardry, so it is easier for future development, and tracking, and releases, to have it all in one repository. This also makes it so that people have only one repository to watch or star and so forth instead of two.